refactor(tests): modularize conftest.py into focused fixture modules
Split monolithic conftest.py into specialized fixture modules for improved maintainability and test organization. Enhanced workflow fixtures with more robust environment detection and higher fidelity git repo initialization.
