+++ tests/fixtures/__init__.py
@@ -1,0 +1,5 @@
+"""Centralized, importable test fixtures package.
+
+Loaded via `pytest_plugins` from the root `tests/conftest.py` to avoid a
+monolithic conftest and improve discoverability.
+"""
